The sender was bypassing the channel. Now it is fixed.
Due to a mistake in the code the sender was not setting the destination
address and port of the channel in the datagrams it sent.
Therefore the default values set by HipsterPacket were used and packets went
directly to the receiver.
As a bonus (christmas gift?) HipsterPacket has been updated in order not to
set any address and port into the datagram thus raising an exception if they
are not manually set before sending the packet.