Now the sender is able to adapt its window.
It retransmits on a dupACK but it does not estimate the RTT. It uses instead
a timer to pause every time the window is full.
On a dupACK the window is halved. On a succesful transmission the window is
reset to WINDOW_SIZE. Such behaviour is a bit aggressive but it should work.